Fixed Bugs in viewing the Sets in My-Flashcards also separated the Sample Popular FlashCards from Dashboard.tsx
